requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
filelock==3.15.4
filetype==1.2.0
Flask==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.7.0
fsspec==2024.6.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
jsonpath-python==1.0.6
kiwisolver==1.4.5
langdetect==1.0.9
layoutparser==0.3.4
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.5.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
openai==1.91.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orderly-set==5.2.2
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.18.0
pikepdf==9.2.0
pillow==10.4.0
portalocker==2.10.1
propcache==0.3.2
proto-plus==1.24.0
protobuf==5.28.0
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pypandoc==1.13
pyparsing==3.1.4
pypdf==4.3.1
pypdfium2==4.30.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.1
pywin32==306
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.4
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
tabulate==0.9.0
timm==1.0.9
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.2
twilio==9.6.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.9
unstructured-client==0.25.6
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==2.2.2
Werkzeug==3.1.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.20.1
